refactor: lift sumCaseUplcConstrImpl onto SumCaseUplcConstrSirTypeGenerator (Phase 5)
Move the @UplcRepr(UplcConstr) sum-source conversion path out of
SumDispatch.sumCaseUplcConstrImpl into
SumCaseUplcConstrSirTypeGenerator.emitConvert. Self-recursion now uses
emitConvert, and the typegen-self default-repr lookups inline (no
fully-qualified SumCaseUplcConstrSirTypeGenerator.foo references).

SumDispatch.sumCaseImpl is bumped to package-private[lowering] so the
lifted emitConvert can delegate the
(DataConstr/PairIntDataList → SumBuiltinList/PackedSumDataList/PairIntDataList)
arm there. SumDispatch.toRepresentation routes
`case SumCaseUplcConstrSirTypeGenerator => SumCaseUplcConstrSirTypeGenerator.emitConvert(...)`.
Net SumDispatch shrink: ~70 lines.
